test for update_message initiated
